+++ include/navigation/transforms.h
@@ -8,10 +8,9 @@
 #ifndef INCLUDE_NAVIGATION_TRANSFORMS_H_
 #define INCLUDE_NAVIGATION_TRANSFORMS_H_
 
-#include "navigation/constants.h"
 #include "Eigen/Core"
 #include "Eigen/Dense"
-#include "global_defs/global_defs.h"
+#include "units/units.h"
 
 namespace navigation {
 /* Euler (3-2-1) to Direction Cosine Matrix (DCM) */
@@ -114,20 +113,20 @@ Eigen::Vector3d ned2lla(const Eigen::Vector3d &loc, const Eigen::Vector3d &ref);
 /* Converts a +/- 180 value to a 0 - 360 value */
 template<typename T>
 T Constrain2Pi(T ang) {
-  ang = std::fmod(ang, static_cast<T>(2) * global::constants::PI<T>);
+  ang = std::fmod(ang, static_cast<T>(2) * constants::PI<T>);
   if (ang < static_cast<T>(0)) {
-    ang += static_cast<T>(2) * global::constants::PI<T>;
+    ang += static_cast<T>(2) * constants::PI<T>;
   }
   return ang;
 }
 /* Converts a 0 - 360 value to a +/- 180 value */
 template<typename T>
 T ConstrainPi(T ang) {
-  if (ang > global::constants::PI<T>) {
-    ang -= static_cast<T>(2) * global::constants::PI<T>;
+  if (ang > constants::PI<T>) {
+    ang -= static_cast<T>(2) * constants::PI<T>;
   }
-  if (ang < -global::constants::PI<T>) {
-    ang += static_cast<T>(2) * global::constants::PI<T>;
+  if (ang < -constants::PI<T>) {
+    ang += static_cast<T>(2) * constants::PI<T>;
   }
   return ang;
 }